src/main/java/com/interview/leetcode/Leetcode_1_TwoSum.java
package com.interview.leetcode;

import java.util.HashMap;
import java.util.Map;

/**
 * https://leetcode.com/problems/two-sum/
 * Difficulty level: Easy
 * Language: Java
 * Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
 * You may assume that each input would have exactly one solution, and you may not use the same element twice.
 * You can return the answer in any order.
 * Example 1:
 * Input: nums = [2,7,11,15], target = 9
 * Output: [0,1]
 * Explanation: Because nums[0] + nums[1] == 9, we return [0, 1].
 * Example 2:
 * Input: nums = [3,2,4], target = 6
 * Output: [1,2]
 * Example 3:
 * Input: nums = [3,3], target = 6
 * Output: [0,1]
 * Constraints:
 * 2 <= nums.length <= 104
 * -109 <= nums[i] <= 109
 * -109 <= target <= 109
 * Only one valid answer exists.
 */
public class Leetcode_1_TwoSum {
    public static void main(String[] args) {
        int[] numbers = new int[]{2, 3, 7, 4, 8};
        ;
        int target = 19;
        int[] resultSolution1 = twoSumSolution1(numbers, target);
        int[] resultSolution2 = twoSumSolution2(numbers, target);
        System.out.println("[" + resultSolution1[0] + "," + resultSolution1[1] + "]");
        System.out.println("[" + resultSolution2[0] + "," + resultSolution2[1] + "]");
    }

    public static int[] twoSumSolution1(int[] numbers, int target) {
        Map<Integer, Integer> visitedNumbers = new HashMap<>();
        for (int i = 0; i < numbers.length; i++) {
            int delta = target - numbers[i];
            if (visitedNumbers.containsKey(delta)) {
                return new int[]{i, visitedNumbers.get(delta)};
            }
            visitedNumbers.put(numbers[i], i);
        }
        return new int[]{-1, -1};
    }

    public static int[] twoSumSolution2(int[] nums, int target) {
        int n = nums.length;
        Map<Integer, Integer> map = new HashMap<>();
        int[] result = new int[2];
        for (int i = 0; i < n; i++) {
            if (map.containsKey(target - nums[i])) {
                result[1] = i;
                result[0] = map.get(target - nums[i]);
                return result;
            }
            map.put(nums[i], i);

        }
        return result;
    }
}
